HAL-1177: move prepared transaction management to runtime
