+++ gui/src/main/java/org/jboss/as/console/client/preview/content/messaging_provider.html
@@ -13,9 +13,6 @@
     <p>Define broadcast and discovery groups, which control how connectors are handled and distributed by this provider.</p>
     <p>Manage cluster connections to load balance between nodes of a cluster.</p>
 
-    <h3>Prepared Transactions</h3>
-    <p>Commit or rollback prepared transactions on this provider.</p>
-
     <h3>Provider settings</h3>
     <p>Configure the settings of this messaging provider.</p>
 </div>
