+++ gui/src/main/java/org/jboss/as/console/client/preview/content/messaging_provider_ja.html
@@ -13,9 +13,6 @@
     <p>Define broadcast and discovery groups which control how connectors are handled and distributed by this provider.</p>
     <p>Manage cluster connections to group servers into clusters so that messages can be load balanced between the nodes of the cluster.</p>
 
-    <h3>Prepared Transactions</h3>
-    <p>Commit or rollback prepared transactions on this provider.</p>
-
     <h3>Provider settings</h3>
     <p>Fine tune the settings of this messaging provider.</p>
 </div>
